added ability to turn on and off particular slack notifications
You will have to update your config file for this to work. This is why I made a branch. Also lets start using branches since we are for sure going to be making changes that the other person may not fully agree with originally (AKA me), but then after a while may see the benefit of the feature.